# docs/swagger.yml
swagger: '2.0'
info:
  description: 'This is documentation for Tech Community - A Social platform for ICT.'
  version: '1.0.0'
  title: 'Tech Community'
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: 'email@email.com'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:3000'
basePath: '/api/v1'
...
schemes:
  - 'https'
  - 'http'
consumes:
  - application/json
  - application/xml
produces:
  - application/json
  - application/xml
  - text/html
securityDefinitions:
  bearerAuth:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /user:
    get:
      summary: 'Get the current user'
      security:
        - bearerAuth: []
      tags:
        - 'Users'
      operationId: getUser
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
      responses:
        '200':
          description: return current user
          schema:
            $ref: '#/definitions/user'

    put:
      summary: 'Update user profile'
      security:
        - bearerAuth: []
      tags:
        - 'Users'
      operationId: updateUser
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/updateUser'
      responses:
        '200':
          description: returns user information
          schema:
            type: object
            properties:
              user:
                $ref: '#/definitions/user'
...
  /profiles/{username}:
    put:
      summary: 'Update user profile'
      security:
        - bearerAuth: []
      tags:
        - 'Profiles'
      operationId: updateProfile
      parameters:
        - name: username
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/updateUser'
      responses:
        '200':
          description: return update user information
          schema:
            $ref: '#/definitions/profile'
    get:
      summary: Get profile by username
      tags:
        - 'Profiles'
      operationId: getProfile
      parameters:
        - name: username
          in: path
          required: true
          type: string
      responses:
        '200':
          description: return a profile
          schema:
            type: object
            properties:
              status:
                type: number
              message:
                type: string
              profile:
                $ref: '#/definitions/profile'
...
  /posts:
    get:
      summary: 'Fetch all the posts'
      tags:
        - 'Posts'
      operationId: getPosts
      responses:
        '200':
          description: returns multiple posts
          schema:
            $ref: '#/definitions/multiplePosts'
    post:
      summary: 'Create a new post'
      security:
        - bearerAuth: []
      tags:
        - 'Posts'
      description: add a new parcel
      operationId: createPost
      parameters:
        - name: Authorization
          in: header
          description: JWT Token
          required: true
          type: string
        - in: body
          name: body
          schema:
            $ref: '#/definitions/createPost'
      responses:
        '201':
          description: returns the created post
          schema:
            type: object
            properties:
              success:
                type: boolean
              message:
                type: string
              data:
                $ref: '#/definitions/post'
...
  /posts/{slug}:
    get:
      summary: 'Fetch a single post'
      tags:
        - 'Posts'
      parameters:
        - in: path
          name: slug
          required: true
          type: string
      responses:
        '200':
          description: returns a single post
          schema:
            $ref: '#/definitions/post'
    put:
      tags:
        - 'Posts'
      summary: 'Update a specific post'
      security:
        - bearerAuth: []
      operationId: updatePost
      parameters:
        - name: Authorization
          in: header
          description: JWT Token
          required: true
          type: string
        - name: slug
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/updatePost'
      responses:
        '200':
          description: returns an updated post
          schema:
            $ref: '#/definitions/post'
    delete:
      tags:
        - 'Posts'
      summary: 'Delete a specific post'
      security:
        - bearerAuth: []
      operationId: deletePost
      parameters:
        - name: Authorization
          in: header
          description: JWT Token
          required: true
          type: string
        - name: slug
          in: path
          required: true
          type: string
      responses:
        '200':
          description: returns a delete post
          schema:
            $ref: '#/definitions/post'
...
  /posts/{postSlug}/like:
    post:
      tags:
        - 'Posts'
      summary: Like a specific post
      security:
        - bearerAuth: []
      operationId: likePost
      parameters:
        - name: Authorization
          in: header
          description: JWT Token
          required: true
          type: string
        - name: slug
          in: path
          required: true
          type: string
      responses:
        '200':
          description: return an post
          schema:
            $ref: '#/definitions/post'
...
definitions:
  user:
    type: object
    properties:
      id:
        type: string
        format: uuid
      email:
        type: string
        format: email
      username:
        type: string
      firstName:
        type: string
        example: 'Jake'
      lastName:
        type: string
        example: 'Jake'
      gender:
        type: string
      birthDate:
        type: string
        example: '2019-01-07'
      bio:
        type: string
      image:
        type: string
      cover:
        type: string
      status:
        type: string
        example: 'active'
      token:
        type: string
    required:
      - id
      - email
      - username
      - token
...
  post:
    type: object
    properties:
      post:
        type: object
        properties:
          slug:
            type: string
          title:
            type: string
          description:
            type: string
          body:
            type: string
          tagList:
            type: array
            items:
              type: string
          createdAt:
            type: string
          updatedAt:
            type: string
          favorited:
            type: boolean
          favoriteCount:
            type: number
          author:
            type: object
            properties:
              username:
                type: string
              bio:
                type: string
              image:
                type: string
              following:
                type: boolean
  multiplePosts:
    type: object
    properties:
      posts:
        type: array
        items:
          type: object
      postsCount:
        type: number
  profile:
    type: object
    properties:
      username:
        type: string
      bio:
        type: string
      image:
        type: string
      following:
        type: boolean
...
  updateUser:
    type: object
    properties:
      user:
        type: object
        properties:
          email:
            type: string
          bio:
            type: string
          image:
            type: string
  createPost:
    type: object
    properties:
      post:
        type: object
        properties:
          title:
            type: string
          description:
            type: string
          body:
            type: string
          tagList:
            type: array
            items:
              type: string
    required:
      - title
      - description
      - body
  updatePost:
    type: object
    properties:
      post:
        type: object
        properties:
          title:
            type: string
...